issues 10

+++ src/main/java/dev/jef/CadastroDeNinjas/missoes/MissoesService.java
@@ -3,6 +3,8 @@ package dev.jef.CadastroDeNinjas.missoes;
 
 import dev.jef.CadastroDeNinjas.ninjas.NinjaModel;
 import dev.jef.CadastroDeNinjas.ninjas.NinjaRepository;
+import exceptions.MissaoNaoEncontradaException;
+import exceptions.NinjaNaoEncontradoException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -22,8 +24,8 @@ public class MissoesService {
 
     // 2. Buscar uma missao por ID
     public MissoesModel buscarMissaoPorId(Long id) {
-        Optional<MissoesModel> missao = missoesRepository.findById(id);
-        return missao.orElse(null); // Retorna a missao se achar, ou null se não encontrar
+        return missoesRepository.findById(id)
+                .orElseThrow(() -> new MissaoNaoEncontradaException(id));
     }
 
     // 3. Criar / Salvar uma nova missão
@@ -33,7 +35,21 @@ public class MissoesService {
 
     // 4. Deletar um ninja por ID
     public void deletarMissaoPorId(Long id) {
+        if (!missoesRepository.existsById(id)) {
+            throw new MissaoNaoEncontradaException(id);
+        }
         missoesRepository.deleteById(id);
+    }
+
+    // 5. Atualizar / Alterar dados da missao
+    public MissoesModel atualizarMissao(Long id, MissoesModel missaoAtualizada) {
+        MissoesModel missaoExistente = missoesRepository.findById(id)
+                .orElseThrow(() -> new MissaoNaoEncontradaException(id));
+
+        missaoExistente.setNome(missaoAtualizada.getNome());
+        missaoExistente.setDificuldade(missaoAtualizada.getDificuldade());
+
+        return missoesRepository.save(missaoExistente);
     }
